Add independence check and util header file for hmm tests.
